Move charmap files, fix makefile about recursion
