+++ SAUCE/defines/charmap/bgm_charmap.h
