+++ SAUCE/defines/charmap/color1_charmap.h
