+++ SAUCE/defines/charmap/luckydraw_charmap.h
